% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim.R
\name{asim1}
\alias{asim1}
\title{Abundance swap}
\usage{
asim1(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by reshuffling populations equiprobably among the nonempty cells
of the entire matrix.
}
\details{
Performs an abundance swap equivalent to algorithm PM in Ulrich and Gotelli (2010), and to the
Fortran `subroutine pm` in Ulrich (2008).

Preserves total matrix abundance and occurrences, but allows marginal sums (i.e. species and
site abundances) to vary equiprobably.
}
\examples{
amat <- rmat(20,12) # creates a random 20×12 abundance matrix
asim1(amat)         # performs equiprobable abundance swap algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.
Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. Ecology 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim11}()},
\code{\link{asim12}()},
\code{\link{asim13}()},
\code{\link{asim14}()},
\code{\link{asim2}()},
\code{\link{asim3}()},
\code{\link{asim4}()},
\code{\link{asim5}()},
\code{\link{asim6}()},
\code{\link{asim7}()},
\code{\link{asim8}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
