% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim_f.R
\name{asim10_f}
\alias{asim10_f}
\title{Resampling with row/column abundances fixed}
\usage{
asim10_f(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by assigning individuals to matrix cells with probabilities
proportional to observed row and column abundance totals until, for each row and column,
total abundances are reached.
}
\details{
Resamples species abundances conditional on and constrained to row and column sums and
constrained to observed species richness; equivalent to algorithm IT in Ulrich and Gotelli
(2010), and to the Fortran \verb{subroutine itt} in Ulrich (2008).

Preserves species and site abundances, but allows occurrences to vary.
}
\examples{
amat <- rmat(20,12) # creates a random 20 x 12 abundance matrix
asim10_f(amat)      # performs proportional abundance resampling algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.

Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. \emph{Ecology} 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim11_f}()},
\code{\link{asim11}()},
\code{\link{asim12_f}()},
\code{\link{asim12}()},
\code{\link{asim13_f}()},
\code{\link{asim13}()},
\code{\link{asim14_f}()},
\code{\link{asim14}()},
\code{\link{asim1_f}()},
\code{\link{asim1}()},
\code{\link{asim2_f}()},
\code{\link{asim2}()},
\code{\link{asim3_f}()},
\code{\link{asim3}()},
\code{\link{asim4_f}()},
\code{\link{asim4}()},
\code{\link{asim5_f}()},
\code{\link{asim5}()},
\code{\link{asim6_f}()},
\code{\link{asim6}()},
\code{\link{asim7_f}()},
\code{\link{asim7}()},
\code{\link{asim8_f}()},
\code{\link{asim8}()},
\code{\link{asim9_f}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
