% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim.R
\name{asim11}
\alias{asim11}
\title{Resampling species by species with row abundances fixed}
\usage{
asim11(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by sequentially (row after row) assigning individuals to each
row with probabilities proportional to observed column abundance totals until the respective
total row abundance is reached.
}
\details{
Resamples species abundances conditional on column sums and constrained to observed row abundance;
equivalent to algorithm ITR in Ulrich and Gotelli (2010), and to the Fortran `subroutine itr` in
Ulrich (2008).
}
\examples{
amat <- rmat(20,12) # creates a random 20 x 12 abundance matrix
asim11(amat)        # performs proportional abundance resampling algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.

Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. Ecology 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim12}()},
\code{\link{asim13}()},
\code{\link{asim14}()},
\code{\link{asim1}()},
\code{\link{asim2}()},
\code{\link{asim3}()},
\code{\link{asim4}()},
\code{\link{asim5}()},
\code{\link{asim6}()},
\code{\link{asim7}()},
\code{\link{asim8}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
