% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim.R
\name{asim13}
\alias{asim13}
\title{Resampling with total abundance fixed}
\usage{
asim13(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by assigning all individuals to matrix cells with probabilities
proportional to observed row and column abundance totals until the matrix-wide total abundance
is reached.
}
\details{
Resamples species abundances conditional on row and column sums and constrained to observed total
abundance; equivalent to algorithm IA in Ulrich and Gotelli (2010), and to the Fortran `subroutine ia`
in Ulrich (2008).
}
\section{Caution}{

May generate matrices with empty rows(species) or columns (sites).
}

\examples{
amat <- rmat(20,12) # creates a random 20 x 12 abundance matrix
asim13(amat)        # performs proportional abundance resampling algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.

Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. Ecology 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim11}()},
\code{\link{asim12}()},
\code{\link{asim14}()},
\code{\link{asim1}()},
\code{\link{asim2}()},
\code{\link{asim3}()},
\code{\link{asim4}()},
\code{\link{asim5}()},
\code{\link{asim6}()},
\code{\link{asim7}()},
\code{\link{asim8}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
