% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim_f.R
\name{asim14_f}
\alias{asim14_f}
\title{Resampling with row/column species richness and abundances fixed}
\usage{
asim14_f(speciesData, swaps = NULL)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Two-step algorithm for randomizing an abundance matrix.
}
\details{
Equivalent to algorithm IF in Ulrich and Gotelli (2010), and to the Fortran \verb{subroutine iff} in
Gotelli (2008).
}
\section{Caution}{

May generate matrices with empty rows(species) or columns (sites).
}

\examples{
amat <- rmat(20,12) # creates a random 20 x 12 abundance matrix
asim14_f(amat)      # performs proportional abundance resampling algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.

Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. \emph{Ecology} 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10_f}()},
\code{\link{asim10}()},
\code{\link{asim11_f}()},
\code{\link{asim11}()},
\code{\link{asim12_f}()},
\code{\link{asim12}()},
\code{\link{asim13_f}()},
\code{\link{asim13}()},
\code{\link{asim14}()},
\code{\link{asim1_f}()},
\code{\link{asim1}()},
\code{\link{asim2_f}()},
\code{\link{asim2}()},
\code{\link{asim3_f}()},
\code{\link{asim3}()},
\code{\link{asim4_f}()},
\code{\link{asim4}()},
\code{\link{asim5_f}()},
\code{\link{asim5}()},
\code{\link{asim6_f}()},
\code{\link{asim6}()},
\code{\link{asim7_f}()},
\code{\link{asim7}()},
\code{\link{asim8_f}()},
\code{\link{asim8}()},
\code{\link{asim9_f}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
