% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asimTest.R
\name{asimTest}
\alias{asimTest}
\title{Tool for testing randomization algorithms}
\usage{
asimTest(func, tmat = NULL, r = 10, ...)
}
\arguments{
\item{func}{[asim*()] function to test (passed as symbol)}

\item{tmat}{test matrix; default option generates a random matrix using [rmat()]}

\item{r}{number of randomizations to perform}
}
\description{
Test for preservation of various matrix metrics by abundance matrix randomization algorithms
(null models)
}
\details{

}
\examples{
asimTest(asim10, r=100)
}
\seealso{
[rmat()], [summarize.mat()]
}
