% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmat.R
\name{rmat}
\alias{rmat}
\title{Random matrix generator}
\usage{
rmat(nr = NULL, nc = NULL, p = NULL, N = NULL)
}
\arguments{
\item{nr}{integer value for number of rows}

\item{nc}{integer value for number of columns}

\item{p}{proportion of cells to be non-zero}

\item{N}{integer value for mean value of non-zero cells}
}
\description{
Create a random abundance matrix, with optional controls for dimensions, probability of occurrence,
and mean cell abundance.
}
\details{

}
\examples{
nr <- 20
nc <- 12
p <- 0.35
N <- 20
amat <- rmat(nr,nc,p,N)   # creates a random 20 x 12 abundance matrix
dim(amat)                 # [1] 20 12 (nr nc)
sum(amat!=0)/length(amat) # converges to 0.35 (p)
mean(amat[amat!=0])       # converges to 20 (N)
}
