% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim.R
\name{asim2}
\alias{asim2}
\title{Abundance swap within rows}
\usage{
asim2(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by reshuffling populations equiprobably among the nonempty cells
of each row.
}
\details{
Performs a series of abundance swaps within rows (species); equivalent to algorithm PR in Ulrich
and Gotelli (2010), and to the Fortran `subroutine pr` in Ulrich (2008).

Preserves total matrix abundance and occurrences as well as species abundances (row totals), but
allows site abundances (column totals) to vary equiprobably.
}
\examples{
amat <- rmat(20,12) # creates a random 20×12 abundance matrix
asim2(amat)         # performs equiprobable row abundance swap algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.
Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. Ecology 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim11}()},
\code{\link{asim12}()},
\code{\link{asim13}()},
\code{\link{asim14}()},
\code{\link{asim1}()},
\code{\link{asim3}()},
\code{\link{asim4}()},
\code{\link{asim5}()},
\code{\link{asim6}()},
\code{\link{asim7}()},
\code{\link{asim8}()},
\code{\link{asim9}()}
}
\concept{abundance matrix null models}
