% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asim.R
\name{asim9}
\alias{asim9}
\title{Resampling site by site with column species richness fixed}
\usage{
asim9(speciesData)
}
\arguments{
\item{speciesData}{species-by-site abundance matrix}
}
\value{
shuffled abundance matrix
}
\description{
Randomizes an abundance matrix by sequentially (column by column) assigning individuals to each
column with probabilities proportional to observed row abundance totals until the respective
column total species richness is reached.
}
\details{
Resamples species abundances conditional on row sums and constrained to observed site occurrences;
equivalent to algorithm ISC in Ulrich and Gotelli (2010), and to the Fortran `subroutine isc` in
Ulrich (2008).

Preserves total occurrences and site richnesses, but allows abundances and species occurrences to
vary.
}
\examples{
amat <- rmat(20,12) # creates a random 20×12 abundance matrix
asim9(amat)         # performs proportional abundance resampling algorithm
}
\references{
Ulrich, W. 2008. Program Co-Occurrence. Version 1.
Ulrich, W. and Gotelli, N.J. 2010. Null model analysis of species associations using abundance data. Ecology 91:3384-3397.
}
\seealso{
Other abundance matrix null models: 
\code{\link{asim10}()},
\code{\link{asim11}()},
\code{\link{asim12}()},
\code{\link{asim13}()},
\code{\link{asim14}()},
\code{\link{asim1}()},
\code{\link{asim2}()},
\code{\link{asim3}()},
\code{\link{asim4}()},
\code{\link{asim5}()},
\code{\link{asim6}()},
\code{\link{asim7}()},
\code{\link{asim8}()}
}
\concept{abundance matrix null models}
