% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asimTest.R
\name{asimTest}
\alias{asimTest}
\title{Tool for testing randomization algorithms}
\usage{
asimTest(func, tmat = NULL, rep = 100, ...)
}
\arguments{
\item{func}{\code{\link[=asim*]{asim*()}} function to test (passed as symbol)}

\item{tmat}{test matrix; default option generates a random matrix using \code{\link[=rmat]{rmat()}}}

\item{...}{additional arguments to be passed to \code{\link[=rmat]{rmat()}}}

\item{r}{number of randomizations to perform}
}
\description{
Test for preservation of various matrix metrics by abundance matrix randomization algorithms
(null models)
}
\details{

}
\examples{
asimTest(asim10, r=100)
}
\seealso{
\code{\link[=rmat]{rmat()}}, \code{\link[=summarize.mat]{summarize.mat()}}
}
